+++ Runtime/Scripts/Utilities/Vector/Vector2Utils.cs
@@ -1,9 +1,9 @@
 /*
- * File Name: VectorExtensions.cs
+ * File Name: Vector2Utils.cs
  * Description: This script is for ...
  * 
  * Author(s): Kokowolo, Will Lacey
- * Date Created: March 18, 2024
+ * Date Created: June 18, 2024
  * 
  * Additional Comments:
  *      File Line Length: 140
@@ -15,19 +15,14 @@ using UnityEngine;
 
 namespace Kokowolo.Utilities
 {
-    public static class VectorExtensions
+    public static class Vector2Utils
     {
         /************************************************************/
         #region Functions
 
-        public static Vector3Int ToVector3Int(this Vector3 v)
+        public static Vector3 GetDirection(Vector3 from, Vector3 to)
         {
-            return new Vector3Int((int) v.x, (int) v.y, (int) v.z);
-        }
-
-        public static Vector3 ToVector3(this Vector2 v)
-        {
-            return new Vector3(v.x, v.y);
+            return to - from;
         }
 
         #endregion